add: routes to the app
