+++ the-experts-fe/package.json
@@ -5,9 +5,9 @@
   "dependencies": {
     "bootstrap": "^4.4.0",
     "env2": "^2.2.2",
-    "joi": "^14.3.1",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "0.9.5"
   },
   "devDependencies": {},
