+++ the-experts-fe/src/components/NavBar/NavBar.css
@@ -1,45 +1,34 @@
-.topnav {
-  overflow: hidden;
-  background-color: #333;
+#navbar {
+  background: black;
+  padding: 0.5em 1em;
+  position: sticky;
+  top: 0;
+  z-index: 1;
+}
+
+#navbar .nav-items {
   display: flex;
-  justify-content: space-between;
+  justify-content: start;
+
 }
 
-.topnav a {
-  float: left;
-  display: block;
-  color: #f2f2f2;
-  text-align: center;
-  padding: 14px 16px;
-  text-decoration: none;
-  font-size: 25px;
-  font-weight: bold;
-}
-
-
-.topnav .icon {
-  display: none;
-}
-
-
-
-.topnav a:hover {
-  background-color: #555;
+.inactive {
   color: white;
+  text-decoration: none;
+  margin-right: 1.5em;
 }
 
+.inactive:hover {
+  color: white;
+  text-decoration: none;
+}
 
-@media screen and (max-width: 400px) {
-  .topnav a {
-    float: left;
-    display: block;
-    color: #f2f2f2;
-    text-align: center;
-    padding: 14px 16px;
-    text-decoration: none;
-    font-size: 15px;
-    font-weight: bold;
-  }
+.active {
+  color: yellow;
+  text-decoration: none;
+}
 
-
+.active:hover {
+  color: yellow;
+  text-decoration: none;
 }
