+++ the-experts-fe/src/components/NavBar/NavBar.jsx
@@ -1,21 +1,35 @@
-
 import React from "react";
+import { Route, NavLink, Switch } from "react-router-dom";
 
 import "./NavBar.css";
 
-export default function NavBar({firstParm,secondParam}) {
+export default function NavBar({ firstParm, secondParam }) {
   return (
     <React.Fragment>
+      <nav id="navbar">
+        <div className="nav-items">
+          <NavLink exact className="inactive" activeClassName="active" to="/">
+            The Experts
+          </NavLink>
+          <NavLink
+            className="inactive"
+            exact
+            activeClassName="active"
+            to="/login"
+          >
+            {firstParm}
+          </NavLink>
 
-<div className="topnav" id="myTopnav">
-  
-  <a href="/" className="login" >{firstParm}</a>
-  <a href="/" className="title" >The Experts</a>
-  <a href="/" className="register" >{secondParam}</a>
-
-</div>
-
-
+          <NavLink
+            exact
+            className="inactive"
+            activeClassName="active"
+            to="/register"
+          >
+            {secondParam}
+          </NavLink>
+        </div>
+      </nav>
     </React.Fragment>
   );
 }
