+++ the-experts-fe/src/components/home/home.css
@@ -1,0 +1,32 @@
+#home-header {
+  background: url("https://cdn.cbtnews.com/wp-content/uploads/2017/10/shutterstock_505405795.jpg") no-repeat center center/cover;
+  height: 350px;
+
+}
+
+#home-header h1 {
+  color: white;
+  font-size: 2em;
+  text-align: center;
+  padding-top: 5em;
+  font-weight: bold;
+}
+
+#home-header span {
+  color: yellow;
+}
+
+#select {
+  text-align: center;
+}
+
+.select-content {
+  padding: 3em 0 2em 0;
+}
+
+#why-us {}
+
+#why-us .why-us-content {
+  display: flex;
+
+}
