+++ the-experts-fe/src/index.js
@@ -3,5 +3,11 @@ import ReactDOM from "react-dom";
 import App from "./App";
 import "./index.css";
 import "bootstrap/dist/css/bootstrap.css";
+import { BrowserRouter } from "react-router-dom";
 
-ReactDOM.render(<App />, document.getElementById("root"));
+ReactDOM.render(
+  <BrowserRouter>
+    <App />
+  </BrowserRouter>,
+  document.getElementById("root")
+);
